backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
supabase-py==2.3.1
markitdown==0.0.1
pydantic==2.5.2
httpx==0.25.2